+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -151,6 +151,9 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 			                     && (log.Encounter == Encounter.Map)
 			                     && string.Compare(log.EvtcVersion, "EVTC20230716", StringComparison.OrdinalIgnoreCase) >= 0,
 				"Add support for fractal scale."),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 10, 0, 1)
+			                     && (log.Encounter == Encounter.BanditTrio),
+				"Slightly improved precision for Bandit Trio success detection when recording player leaves the fort."),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};